move drawing-tool into plugins
