+++ src/lib/logger.ts
@@ -8,7 +8,7 @@ import { InvestigationModelType } from "../models/curriculum/investigation";
 import { ProblemModelType } from "../models/curriculum/problem";
 import { DocumentModelType } from "../models/document/document";
 import { JXGChange } from "../models/tools/geometry/jxg-changes";
-import { DrawingToolChange } from "../models/tools/drawing/drawing-types";
+import { DrawingToolChange } from "../plugins/drawing-tool/model/drawing-types";
 import { ITableChange } from "../models/tools/table/table-change";
 import { ENavTab } from "../models/view/nav-tabs";
 import { DEBUG_LOGGER } from "../lib/debug";
