+++ src/models/stores/problem-configuration.ts
@@ -1,6 +1,6 @@
 import { Instance, SnapshotIn, types } from "mobx-state-tree";
 import { IAuthoredDocumentContent } from "../document/document-content-import";
-import { StampModel } from "../tools/drawing/stamp";
+import { StampModel } from "../../plugins/drawing-tool/model/stamp";
 import { ToolButtonModel } from "../tools/tool-button";
 import { SettingsMstType } from "./settings";
