+++ src/plugins/drawing-tool/components/color-swatch.tsx
