+++ src/plugins/drawing-tool/components/drawing-tool.sass
