+++ src/plugins/drawing-tool/components/drawing-tool.tsx
@@ -1,12 +1,12 @@
 import classNames from "classnames";
 import React, { useEffect } from "react";
-import { IToolTileProps } from "../tool-tile";
+import { IToolTileProps } from "../../../components/tools/tool-tile";
 import { ToolbarView } from "./drawing-toolbar";
-import { DrawingLayerView } from "./drawing-layer";
-import { useToolbarToolApi } from "../hooks/use-toolbar-tool-api";
-import { DrawingContentModelType } from "../../../models/tools/drawing/drawing-content";
+import { DrawingLayerView } from "../../../components/tools/drawing-tool/drawing-layer";
+import { useToolbarToolApi } from "../../../components/tools/hooks/use-toolbar-tool-api";
+import { DrawingContentModelType } from "../model/drawing-content";
 import { useCurrent } from "../../../hooks/use-current";
-import { exportDrawingTileSpec } from "../../../models/tools/drawing/drawing-export";
+import { exportDrawingTileSpec } from "../model/drawing-export";
 import { ITileExportOptions } from "../../../models/tools/tool-content-info";
 
 import "./drawing-tool.sass";
