+++ src/plugins/drawing-tool/components/fill-color-palette.tsx
