+++ src/plugins/drawing-tool/components/stamps-palette.tsx
@@ -1,6 +1,6 @@
 import classNames from "classnames";
 import React from "react";
-import { StampModelType } from "../../../models/tools/drawing/stamp";
+import { StampModelType } from "../model/stamp";
 import { StampButton } from "./stamp-button";
 
 interface IProps {
