+++ src/plugins/drawing-tool/components/stroke-color-palette.tsx
