+++ src/plugins/drawing-tool/components/variable-dialog.scss
