+++ src/plugins/drawing-tool/drawing-registration.ts
@@ -1,7 +1,7 @@
-import { registerToolContentInfo } from "../tool-content-info";
-import { DrawingContentModel, DrawingToolMetadataModel, defaultDrawingContent } from "./drawing-content";
-import { kDrawingToolID, kDrawingDefaultHeight } from "./drawing-types";
-import DrawingToolComponent from "../../../components/tools/drawing-tool/drawing-tool";
+import { registerToolContentInfo } from "../../models/tools/tool-content-info";
+import { DrawingContentModel, DrawingToolMetadataModel, defaultDrawingContent } from "./model/drawing-content";
+import { kDrawingToolID, kDrawingDefaultHeight } from "./model/drawing-types";
+import DrawingToolComponent from "./components/drawing-tool";
 import DrawingToolIcon from "../../../clue/assets/icons/draw-tool.svg";
 
 registerToolContentInfo({
