+++ src/plugins/drawing-tool/model/drawing-import.test.ts
