+++ src/plugins/drawing-tool/model/drawing-import.ts
