+++ src/plugins/drawing-tool/model/drawing-objects.ts
