+++ src/plugins/drawing-tool/model/drawing-types.ts
