+++ src/plugins/drawing-tool/model/stamp.ts
